Finish loading state and error state style.
